Updated modules paths in src/models
