+++ src/models/msg/gamm/msg_create_pool.ts
@@ -1,4 +1,4 @@
-import { Categories } from '../../../types';
+import { Categories } from '../types';
 
 class MsgCreatePool {
     public category: Categories;
