+++ src/models/msg/gamm/msg_exit_pool.ts
@@ -1,4 +1,4 @@
-import { Categories } from '../../../types';
+import { Categories } from '../types';
 
 class MsgExitPool {
     public category: Categories;
