+++ src/models/msg/gamm/msg_exit_swap_share_amount_in.ts
@@ -1,4 +1,4 @@
-import { Categories } from '../../../types';
+import { Categories } from '../types';
 
 class MsgExitSwapShareAmountIn {
     public category: Categories;
