+++ src/models/msg/gamm/msg_join_pool.ts
@@ -1,31 +1,31 @@
-import { Categories } from '../../../types';
+import { Categories } from '../types';
 
-class MsgJoinSwapExternAmountIn {
+class MsgJoinPool {
     public category: Categories;
     public type: string;
     public poolId: string;
     public sender: string;
-    public tokenIn: any;
-    public shareOutMinAmount: any;
+    public tokenInMaxs: any;
+    public shareOutAmount: string;
 
     constructor(payload: any) {
       this.category = 'osmosis';
       this.type = payload.type;
       this.poolId = payload.poolId;
       this.sender = payload.sender;
-      this.tokenIn = payload.tokenIn;
-      this.shareOutMinAmount = payload.shareOutMinAmount;
+      this.tokenInMaxs = payload.tokenInMaxs;
+      this.shareOutAmount = payload.shareOutAmount;
     }
 
     static fromJson(json: any) {
-      return new MsgJoinSwapExternAmountIn({
+      return new MsgJoinPool({
         type: json['@type'],
         poolId: json.poolId,
         sender: json.sender,
-        tokenIn: json.tokenIn,
-        shareOutMinAmount: json.shareOutMinAmount,
+        tokenInMaxs: json.tokenInMaxs,
+        shareOutAmount: json.shareOutAmount,
       });
     }
 }
 
-export default MsgJoinSwapExternAmountIn;
+export default MsgJoinPool;
