+++ src/models/msg/gamm/msg_join_swap_share_amount_out.ts
@@ -1,4 +1,4 @@
-import { Categories } from '../../../types';
+import { Categories } from '../types';
 
 class MsgJoinSwapShareAmountOut {
     public category: Categories;
