+++ src/models/msg/gamm/msg_swap_exact_amount_out.ts
@@ -1,31 +1,31 @@
-import { Categories } from '../../../types';
+import { Categories } from '../types';
 
-class MsgSwapExactAmountIn {
+class MsgSwapExactAmountOut {
     public category: Categories;
     public type: string;
     public routes: any;
     public sender: string;
-    public tokenIn: any;
-    public tokenOutMinAmount: string;
+    public tokenOut: any;
+    public tokenInMaxAmount: any;
 
     constructor(payload: any) {
       this.category = 'osmosis';
       this.type = payload.type;
       this.routes = payload.routes;
       this.sender = payload.sender;
-      this.tokenIn = payload.tokenIn;
-      this.tokenOutMinAmount = payload.tokenOutMinAmount;
+      this.tokenOut = payload.tokenOut;
+      this.tokenInMaxAmount = payload.tokenInMaxAmount;
     }
 
     static fromJson(json: any) {
-      return new MsgSwapExactAmountIn({
+      return new MsgSwapExactAmountOut({
         type: json['@type'],
         routes: json.routes,
         sender: json.sender,
-        tokenIn: json.tokenIn,
-        tokenOutMinAmount: json.tokenOutMinAmount,
+        tokenOut: json.tokenOut,
+        tokenInMaxAmount: json.tokenInMaxAmount,
       });
     }
 }
 
-export default MsgSwapExactAmountIn;
+export default MsgSwapExactAmountOut;
