+++ src/models/msg/incentives/msg_create_gauge.ts
@@ -1,4 +1,4 @@
-import { Categories } from '../../types';
+import { Categories } from '../types';
 
 class MsgCreateGauge {
     public category: Categories;
